Fix triangle rendering issue
Triangle did not appear to be rendering previous to this patch.
Polygon face-relative culling had been set to a clockwise bias.
However the vertices of the triangle that was intended to be displayed
on-screen were ordered counter-clockwise. This meant that the backside
of the triangle was being displayed. Culling caused this backside to be
invisible. Ordering the vertices of the triangle in a clockwise fashion
made the triangle render as desired.